优化项目结构
